+++ apps/api-journeys/src/app/emails/EmailLogo/EmailLogo.tsx
@@ -1,0 +1,28 @@
+import { ReactElement } from 'react'
+
+export function EmailLogo(): ReactElement {
+  return (
+    <svg
+      width="148"
+      height="27"
+      viewBox="0 0 148 27"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M42.2019 16.3574V7.84668C42.2019 7.66895 42.2635 7.51628 42.3865 7.38867C42.5095 7.26107 42.6599 7.19727 42.8377 7.19727H44.8474C45.0069 7.19727 45.1756 7.26107 45.3533 7.38867C45.531 7.51628 45.6769 7.66895 45.7908 7.84668L49.9197 14.1973V7.83301C49.9197 7.65527 49.9812 7.50488 50.1043 7.38184C50.2319 7.25879 50.3845 7.19727 50.5623 7.19727H51.7381C51.9158 7.19727 52.0662 7.25879 52.1892 7.38184C52.3168 7.50488 52.3806 7.65983 52.3806 7.84668V16.3574C52.3806 16.5352 52.3168 16.6878 52.1892 16.8154C52.0662 16.9385 51.9158 17 51.7381 17H49.4685C49.2908 17 49.1085 16.9385 48.9217 16.8154C48.7348 16.6878 48.5867 16.5352 48.4773 16.3574L44.697 10.4922V16.3574C44.697 16.5352 44.6332 16.6878 44.5056 16.8154C44.3826 16.9385 44.2322 17 44.0545 17H42.8513C42.6736 17 42.5209 16.9385 42.3933 16.8154C42.2657 16.6878 42.2019 16.5352 42.2019 16.3574ZM55.4926 16.3438V7.83301C55.4926 7.65527 55.5564 7.50488 55.684 7.38184C55.8116 7.25879 55.9642 7.19727 56.142 7.19727H64.0238C64.2015 7.19727 64.3519 7.26107 64.475 7.38867C64.6026 7.51628 64.6664 7.66895 64.6664 7.84668V8.48926C64.6664 8.66699 64.6026 8.81966 64.475 8.94727C64.3519 9.07031 64.2015 9.13184 64.0238 9.13184H58.0697V11.1211H63.1967C63.3744 11.1211 63.5248 11.1849 63.6478 11.3125C63.7754 11.4355 63.8392 11.5859 63.8392 11.7637V12.4062C63.8392 12.584 63.7754 12.7367 63.6478 12.8643C63.5248 12.9873 63.3744 13.0488 63.1967 13.0488H58.0697V15.0654H64.0101C64.1833 15.0654 64.3314 15.127 64.4545 15.25C64.5775 15.373 64.639 15.5234 64.639 15.7012V16.3574C64.639 16.5352 64.5752 16.6878 64.4476 16.8154C64.32 16.9385 64.1674 17 63.9896 17H56.142C55.9642 17 55.8116 16.9385 55.684 16.8154C55.5564 16.6878 55.4926 16.5306 55.4926 16.3438ZM66.6162 17C66.3883 17 66.2607 16.8975 66.2334 16.6924C66.2334 16.5921 66.2812 16.4805 66.3769 16.3574L69.8222 11.9619L66.5752 7.84668C66.4795 7.72819 66.4316 7.61654 66.4316 7.51172C66.4316 7.30208 66.5569 7.19727 66.8076 7.19727H68.2773C68.733 7.19727 69.1318 7.41374 69.4736 7.84668L71.3467 10.2256L73.2197 7.84668C73.5569 7.41829 73.9557 7.20182 74.416 7.19727H75.8857C76.1364 7.19727 76.2617 7.30208 76.2617 7.51172C76.2617 7.61198 76.2138 7.72363 76.1181 7.84668L72.8711 11.9619L76.3164 16.3574C76.4121 16.4759 76.4599 16.5716 76.4599 16.6445C76.4599 16.7174 76.4508 16.7699 76.4326 16.8018C76.3825 16.9339 76.264 17 76.0771 17H74.6211C74.1653 17 73.7689 16.7858 73.4316 16.3574L71.3467 13.6982L69.2617 16.3574C68.9245 16.7858 68.528 17 68.0722 17H66.6162ZM77.8834 8.48926V7.84668C77.8834 7.65983 77.9449 7.50488 78.0679 7.38184C78.191 7.25879 78.3459 7.19727 78.5328 7.19727H87.2076C87.3944 7.19727 87.5494 7.25879 87.6724 7.38184C87.7955 7.50488 87.857 7.65983 87.857 7.84668V8.48926C87.857 8.67611 87.7955 8.83105 87.6724 8.9541C87.5494 9.07259 87.3944 9.13184 87.2076 9.13184H84.1588V16.3711C84.1588 16.5488 84.095 16.6992 83.9674 16.8223C83.8443 16.9408 83.6939 17 83.5162 17H82.2242C82.0465 17 81.8938 16.9408 81.7662 16.8223C81.6431 16.6992 81.5816 16.5488 81.5816 16.3711V9.13184H78.5328C78.3459 9.13184 78.191 9.07259 78.0679 8.9541C77.9449 8.83105 77.8834 8.67611 77.8834 8.48926ZM89.4308 14.4912C89.4308 14.2223 89.5106 14.0309 89.6701 13.917C89.8296 13.7985 89.9845 13.7393 90.1349 13.7393H90.9484C91.3586 13.7393 91.6707 13.9626 91.8849 14.4092C92.0991 14.8558 92.5799 15.0791 93.3273 15.0791L95.7199 15.127C96.522 15.127 96.923 14.8649 96.923 14.3408C96.923 13.9443 96.4331 13.6322 95.4533 13.4043C95.0158 13.2995 94.5327 13.1969 94.0041 13.0967C93.4754 12.9919 92.9491 12.862 92.425 12.707C91.9054 12.5521 91.4269 12.3698 90.9894 12.1602C90.5519 11.9505 90.1965 11.6634 89.923 11.2988C89.6541 10.9297 89.5197 10.4967 89.5197 10C89.5197 9.08398 89.8455 8.37988 90.4972 7.8877C91.1489 7.39095 92.27 7.14258 93.8605 7.14258L95.5968 7.18359C96.5311 7.18359 97.3719 7.42513 98.1193 7.9082C98.8713 8.38672 99.2472 8.986 99.2472 9.70605C99.2472 9.97493 99.1675 10.1686 99.008 10.2871C98.8485 10.4056 98.6958 10.4648 98.55 10.4648H97.7365C97.3263 10.4648 97.0142 10.2415 96.8 9.79492C96.5858 9.34375 96.105 9.11816 95.3576 9.11816L93.3 9.07715C92.4979 9.07715 92.0968 9.33919 92.0968 9.86328C92.0968 10.1504 92.3498 10.3851 92.8556 10.5674C93.3661 10.7497 93.9813 10.9115 94.7013 11.0527C95.4259 11.1895 96.1528 11.3558 96.882 11.5518C97.6112 11.7477 98.2287 12.0713 98.7345 12.5225C99.245 12.9691 99.5002 13.5273 99.5002 14.1973C99.5002 15.1133 99.172 15.8197 98.5158 16.3164C97.8641 16.8086 96.7453 17.0547 95.1593 17.0547L93.0881 17.0137C92.1538 17.0137 91.3107 16.7744 90.5588 16.2959C89.8068 15.8128 89.4308 15.2113 89.4308 14.4912ZM101.006 8.48926V7.84668C101.006 7.65983 101.067 7.50488 101.19 7.38184C101.313 7.25879 101.468 7.19727 101.655 7.19727H110.33C110.517 7.19727 110.672 7.25879 110.795 7.38184C110.918 7.50488 110.979 7.65983 110.979 7.84668V8.48926C110.979 8.67611 110.918 8.83105 110.795 8.9541C110.672 9.07259 110.517 9.13184 110.33 9.13184H107.281V16.3711C107.281 16.5488 107.217 16.6992 107.09 16.8223C106.967 16.9408 106.816 17 106.638 17H105.346C105.169 17 105.016 16.9408 104.888 16.8223C104.765 16.6992 104.704 16.5488 104.704 16.3711V9.13184H101.655C101.468 9.13184 101.313 9.07259 101.19 8.9541C101.067 8.83105 101.006 8.67611 101.006 8.48926ZM113.25 16.3438V7.83301C113.25 7.65527 113.314 7.50488 113.442 7.38184C113.569 7.25879 113.722 7.19727 113.9 7.19727H121.782C121.959 7.19727 122.11 7.26107 122.233 7.38867C122.36 7.51628 122.424 7.66895 122.424 7.84668V8.48926C122.424 8.66699 122.36 8.81966 122.233 8.94727C122.11 9.07031 121.959 9.13184 121.782 9.13184H115.828V11.1211H120.954C121.132 11.1211 121.283 11.1849 121.406 11.3125C121.533 11.4355 121.597 11.5859 121.597 11.7637V12.4062C121.597 12.584 121.533 12.7367 121.406 12.8643C121.283 12.9873 121.132 13.0488 120.954 13.0488H115.828V15.0654H121.768C121.941 15.0654 122.089 15.127 122.212 15.25C122.335 15.373 122.397 15.5234 122.397 15.7012V16.3574C122.397 16.5352 122.333 16.6878 122.205 16.8154C122.078 16.9385 121.925 17 121.747 17H113.9C113.722 17 113.569 16.9385 113.442 16.8154C113.314 16.6878 113.25 16.5306 113.25 16.3438ZM124.832 16.3574V7.84668C124.832 7.66895 124.894 7.51628 125.017 7.38867C125.14 7.26107 125.29 7.19727 125.468 7.19727H130.684C133.441 7.19727 134.819 8.46875 134.819 11.0117C134.819 13.5501 133.441 14.8193 130.684 14.8193H127.409V16.3574C127.409 16.5352 127.345 16.6878 127.218 16.8154C127.095 16.9385 126.944 17 126.767 17H125.481C125.304 17 125.151 16.9385 125.023 16.8154C124.896 16.6878 124.832 16.5352 124.832 16.3574ZM127.409 12.8848H129.904C130.67 12.8848 131.251 12.7572 131.647 12.502C132.044 12.2422 132.242 11.7454 132.242 11.0117C132.242 10.2734 132.044 9.77669 131.647 9.52148C131.251 9.26172 130.67 9.13184 129.904 9.13184H127.409V12.8848ZM136.673 14.4912C136.673 14.2223 136.753 14.0309 136.913 13.917C137.072 13.7985 137.227 13.7393 137.378 13.7393H138.191C138.601 13.7393 138.913 13.9626 139.128 14.4092C139.342 14.8558 139.822 15.0791 140.57 15.0791L142.962 15.127C143.765 15.127 144.166 14.8649 144.166 14.3408C144.166 13.9443 143.676 13.6322 142.696 13.4043C142.258 13.2995 141.775 13.1969 141.247 13.0967C140.718 12.9919 140.192 12.862 139.668 12.707C139.148 12.5521 138.67 12.3698 138.232 12.1602C137.795 11.9505 137.439 11.6634 137.166 11.2988C136.897 10.9297 136.762 10.4967 136.762 10C136.762 9.08398 137.088 8.37988 137.74 7.8877C138.392 7.39095 139.513 7.14258 141.103 7.14258L142.839 7.18359C143.774 7.18359 144.614 7.42513 145.362 7.9082C146.114 8.38672 146.49 8.986 146.49 9.70605C146.49 9.97493 146.41 10.1686 146.251 10.2871C146.091 10.4056 145.938 10.4648 145.793 10.4648H144.979C144.569 10.4648 144.257 10.2415 144.043 9.79492C143.828 9.34375 143.348 9.11816 142.6 9.11816L140.543 9.07715C139.74 9.07715 139.339 9.33919 139.339 9.86328C139.339 10.1504 139.592 10.3851 140.098 10.5674C140.609 10.7497 141.224 10.9115 141.944 11.0527C142.669 11.1895 143.395 11.3558 144.125 11.5518C144.854 11.7477 145.471 12.0713 145.977 12.5225C146.488 12.9691 146.743 13.5273 146.743 14.1973C146.743 15.1133 146.415 15.8197 145.758 16.3164C145.107 16.8086 143.988 17.0547 142.402 17.0547L140.331 17.0137C139.396 17.0137 138.553 16.7744 137.801 16.2959C137.049 15.8128 136.673 15.2113 136.673 14.4912Z"
+        fill="#26262E"
+      />
+      <path
+        d="M8.99783 4C8.99783 2.89543 9.89326 2 10.9978 2H29.22C30.3246 2 31.22 2.89543 31.22 4V19.5449C31.22 20.945 29.8184 21.9118 28.5096 21.4145L10.2874 14.49C9.51106 14.195 8.99783 13.451 8.99783 12.6205V4Z"
+        fill="#9E2630"
+      />
+      <path
+        d="M22.9978 25C22.9978 26.1046 22.1024 27 20.9978 27H2.7756C1.67103 27 0.775602 26.1046 0.775602 25V9.45509C0.775602 8.05495 2.17722 7.08817 3.48604 7.58552L21.7083 14.51C22.4846 14.805 22.9978 15.549 22.9978 16.3795V25Z"
+        fill="#C52D3A"
+      />
+    </svg>
+  )
+}
+
+export default EmailLogo
