+++ apps/api-journeys/src/app/emails/EmailLogo/index.ts
@@ -1,0 +1,1 @@
+export { EmailLogo } from './EmailLogo.tsx'
