+++ apps/api-journeys/src/app/emails/JourneyInvite/index.ts
@@ -1,0 +1,1 @@
+export { JourneyInviteEmail } from './JourneyInvite.tsx'
