+++ apps/api-journeys/src/app/emails/TeamInvite/index.ts
@@ -1,0 +1,1 @@
+export { TeamInviteEmail } from './TeamInvite.tsx'
