+++ apps/api-journeys/src/app/modules/userTeamInvite/userTeamInvite.service.ts
@@ -5,7 +5,7 @@ import { Queue } from 'bullmq'
 
 import { Team } from '.prisma/api-journeys-client'
 
-import TeamInviteEmail from '../../emails/TeamInvite'
+import TeamInviteEmail from '../../emails/TeamInvite/TeamInvite'
 import { EmailJob } from '../email/email.consumer'
 
 @Injectable()
