+++ apps/api-journeys/tsconfig.json
@@ -1,5 +1,9 @@
 {
   "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "allowImportingTsExtensions": true,
+    "noEmit": true
+  },
   "files": [],
   "include": [],
   "references": [
